requirements.txt
aiosmtplib==2.0.1
anyio==3.6.2
appnope==0.1.3
asttokens==2.2.1
backcall==0.2.0
blinker==1.6
certifi==2022.12.7
cffi==1.15.1
click==8.1.3
cryptography==40.0.1
decorator==5.1.1
Deprecated==1.2.13
dnspython==2.3.0
email-validator==1.3.1
executing==1.2.0
fastapi==0.95.0
fastapi-mail==1.2.7
h11==0.14.0
httpcore==0.16.3
httptools==0.5.0
httpx==0.23.3
idna==3.4
ipdb==0.13.13
ipython==8.12.0
itsdangerous==2.1.2
jedi==0.18.2
Jinja2==3.1.2
jwcrypto==1.4.2
MarkupSafe==2.1.2
matplotlib-inline==0.1.6
orjson==3.8.9
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
prompt-toolkit==3.0.38
ptyprocess==0.7.0
pure-eval==0.2.2
pycparser==2.21
pydantic==1.10.7
Pygments==2.14.0
PyJWT==2.6.0
python-dotenv==1.0.0
python-multipart==0.0.6
PyYAML==6.0
rfc3986==1.5.0
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.8
stack-data==0.6.2
starlette==0.26.1
traitlets==5.9.0
typing_extensions==4.5.0
ujson==5.7.0
uvicorn==0.21.1
uvloop==0.17.0
watchfiles==0.19.0
wcwidth==0.2.6
websockets==11.0
wrapt==1.15.0
